requirements.txt
amqp==2.6.1
av==8.0.3
billiard==3.6.4.0
boto3==1.20.23
botocore==1.23.23
celery==4.4.0
# dataclasses==0.8
ffmpeg-python==0.2.0
future==0.18.2
importlib-metadata==4.8.3
jmespath==0.10.0
kombu==4.6.11
numpy==1.19.5
Pillow==8.4.0
PIMS==0.5
pkg_resources==0.0.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
slicerator==1.0.0
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.1
typing_extensions==4.0.1
urllib3==1.26.7
vine==1.3.0
zipp==3.6.0